requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
attrs==25.3.0
black==25.1.0
certifi==2025.7.14
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
coverage==7.10.1
dataclasses-json==0.6.7
distlib==0.4.0
distro==1.9.0
exceptiongroup==1.3.0
faiss-cpu==1.11.0.post1
fastapi==0.116.1
filelock==3.18.0
frozenlist==1.7.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
identify==2.6.12
idna==3.10
iniconfig==2.1.0
isort==6.0.1
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-core==0.3.72
langchain-openai==0.3.28
langchain-text-splitters==0.3.9
langsmith==0.4.8
marshmallow==3.26.1
multidict==6.6.3
mypy==1.17.0
mypy_extensions==1.1.0
nodeenv==1.9.1
numpy==2.0.2
openai==1.97.1
orjson==3.11.1
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
pre_commit==4.2.0
propcache==0.3.2
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyMuPDF==1.26.3
pypdf==5.8.0
pytest==8.4.1
pytest-cov==6.2.1
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
ruff==0.12.5
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.47.2
tenacity==9.1.2
tiktoken==0.9.0
tomli==2.2.1
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
virtualenv==20.32.0
yarl==1.20.1
zstandard==0.23.0
